+++ src/docker_compose-copy.yml
@@ -1,0 +1,15 @@
+version: '3.8'
+services:
+  postgres-db:
+    image: postgres:15.3
+    restart: always
+    environment:
+      POSTGRES_USER:
+      POSTGRES_DB:
+      POSTGRES_PASSWORD:
+    volumes:
+      - ./postgres:/var/lib/postgresql/data
+
+networks:
+  app-network:
+    driver: bridge
