+++ src/main/java/com/david/jpa_project/controller/advice/ControllerAdvice.java
@@ -1,0 +1,45 @@
+package com.david.jpa_project.controller.advice;
+
+import org.hibernate.exception.ConstraintViolationException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
+import org.springframework.web.bind.MissingRequestHeaderException;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import java.util.Map;
+
+@RestControllerAdvice
+public class ControllerAdvice {
+
+    private static final String KEY_MESSAGE = "message";
+
+    @ExceptionHandler(MethodArgumentNotValidException.class)
+    private ResponseEntity<Map<String, String>> handleValidationErrors(
+            MethodArgumentNotValidException ex
+    ) {
+
+        String error = ex.getBindingResult()
+                .getFieldErrors()
+                .stream().map(FieldError::getDefaultMessage)
+                .findFirst().orElse(ex.getMessage());
+
+        return new ResponseEntity<>(Map.of(KEY_MESSAGE, error), HttpStatus.BAD_REQUEST);
+    }
+
+    @ExceptionHandler(MissingRequestHeaderException.class)
+    private ResponseEntity<Map<String, String>> handleMissingRequestHeaderException(
+            MissingRequestHeaderException ex
+    ){
+        return new ResponseEntity<>(Map.of(KEY_MESSAGE, ex.getMessage()), HttpStatus.BAD_REQUEST);
+    }
+
+    @ExceptionHandler(ConstraintViolationException.class)
+    private ResponseEntity<Map<String, String>> handleConstraintViolationException(
+            ConstraintViolationException ex
+    ){
+        return new ResponseEntity<>(Map.of(KEY_MESSAGE, ex.getMessage()), HttpStatus.BAD_REQUEST);
+    }
+}
