+++ src/main/java/com/david/jpa_project/controller/http/employee/products/ProductsController.java
@@ -1,0 +1,16 @@
+package com.david.jpa_project.controller.http.employee.products;
+
+import com.david.jpa_project.services.employee.products.IProductsService;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/products")
+public class ProductsController{
+
+    private final IProductsService productsService;
+
+    public ProductsController(IProductsService productsService) {
+        this.productsService = productsService;
+    }
+}
