+++ src/main/java/com/david/jpa_project/model/entities/entity/User.java
@@ -1,0 +1,31 @@
+package com.david.jpa_project.model.entities.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+@Table(name = "T_USERS", schema = "demo")
+@Inheritance(strategy = InheritanceType.JOINED)
+public class User {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Basic(optional = false)
+    @Column(length = 50)
+    private String name;
+
+    @Basic(optional = false)
+    @Column(length = 50)
+    private String lastName;
+
+    @Transient
+    private String fullName;
+}
