+++ src/main/java/com/david/jpa_project/model/entities/enums/JobTitle.java
@@ -1,0 +1,15 @@
+package com.david.jpa_project.model.entities.enums;
+
+import lombok.Getter;
+
+@Getter
+public enum JobTitle {
+    MANAGER("MANAGER"),
+    SALESMAN("SALESMAN");
+
+    private final String value;
+
+    JobTitle(String value) {
+        this.value = value;
+    }
+}
