+++ src/main/java/com/david/jpa_project/model/entities/enums/OrderStatus.java
@@ -1,0 +1,18 @@
+package com.david.jpa_project.model.entities.enums;
+
+import lombok.Getter;
+
+@Getter
+public enum OrderStatus {
+
+    PENDING("PENDING"),
+    CANCELLED("CANCELLED"),
+    COMPLETED("COMPLETED"),
+    DELIVERED("DELIVERED");
+
+    private final String value;
+
+    OrderStatus(String value) {
+        this.value = value;
+    }
+}
