--- src/main/java/com/david/jpa_project/model/entity/RefreshToken.java
@@ -1,26 +1,0 @@
-package com.david.jpa_project.model.entity;
-
-import com.david.jpa_project.model.embeddable.AuditData;
-import com.david.jpa_project.model.embeddable.RefreshTokenId;
-import jakarta.persistence.*;
-import lombok.*;
-
-@Getter
-@Setter
-@EqualsAndHashCode
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-@Table(name = "T_REFRESH_ACCESS_TOKEN", schema = "demo")
-public class RefreshToken {
-    @EmbeddedId
-    private RefreshTokenId id;
-
-    @MapsId("accessTokenId")
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "access_token_id", nullable = false)
-    private AccessToken accessToken;
-
-    @Embedded
-    private AuditData auditData;
-}
