--- src/main/java/com/david/jpa_project/model/entity/User.java
@@ -1,31 +1,0 @@
-package com.david.jpa_project.model.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-@Table(name = "T_USERS", schema = "demo")
-@Inheritance(strategy = InheritanceType.JOINED)
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Basic(optional = false)
-    @Column(length = 50)
-    private String name;
-
-    @Basic(optional = false)
-    @Column(length = 50, name = "last_name")
-    private String lastName;
-
-    @Transient
-    private String fullName;
-}
