--- src/main/java/com/david/jpa_project/model/enums/JobTitle.java
@@ -1,15 +1,0 @@
-package com.david.jpa_project.model.enums;
-
-import lombok.Getter;
-
-@Getter
-public enum JobTitle {
-    MANAGER("MANAGER"),
-    SALESMAN("SALESMAN");
-
-    private final String value;
-
-    JobTitle(String value) {
-        this.value = value;
-    }
-}
