+++ src/main/java/com/david/jpa_project/model/repositories/CategoryRepository.java
@@ -1,0 +1,10 @@
+package com.david.jpa_project.model.repositories;
+
+import com.david.jpa_project.model.entities.entity.Category;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+
+}
