+++ src/main/java/com/david/jpa_project/model/repositories/EmployeeRepository.java
@@ -1,0 +1,9 @@
+package com.david.jpa_project.model.repositories;
+
+import com.david.jpa_project.model.entities.entity.Employee;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface EmployeeRepository extends JpaRepository<Employee, Long> {
+}
