+++ src/main/java/com/david/jpa_project/model/repositories/UserRepository.java
@@ -1,0 +1,10 @@
+package com.david.jpa_project.model.repositories;
+
+import com.david.jpa_project.model.entities.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {
+
+}
