+++ src/main/java/com/david/jpa_project/services/customer/addresses/IAddressesService.java
@@ -1,0 +1,4 @@
+package com.david.jpa_project.services.customer.addresses;
+
+public interface IAddressesService {
+}
