+++ src/main/java/com/david/jpa_project/services/customer/orders/IOrdersService.java
@@ -1,0 +1,4 @@
+package com.david.jpa_project.services.customer.orders;
+
+public interface IOrdersService {
+}
