+++ src/main/java/com/david/jpa_project/services/employee/categories/ICategoriesService.java
@@ -1,0 +1,4 @@
+package com.david.jpa_project.services.employee.categories;
+
+public interface ICategoriesService {
+}
