+++ src/main/java/com/david/jpa_project/services/employee/products/IProductsService.java
@@ -1,0 +1,4 @@
+package com.david.jpa_project.services.employee.products;
+
+public interface IProductsService {
+}
